updated code in expense bill data migration module

+++ FinanceDataMigration/src/main/java/org/egov/finance/migration/modules/expensebill/service/BillSubtypeMapping.java
@@ -5,24 +5,24 @@ import java.util.Map;
 
 public class BillSubtypeMapping {
 
-	private static final Map<String, Long> BILL_SUBTYPE_MAP = new HashMap<>();
+	private static final Map<String, Integer> BILL_SUBTYPE_MAP = new HashMap<>();
 
 	static {
 
-		BILL_SUBTYPE_MAP.put("Contingent", 1L);
-		BILL_SUBTYPE_MAP.put("Salary", 2L);
-		BILL_SUBTYPE_MAP.put("Pension", 3L);
-		BILL_SUBTYPE_MAP.put("Works", 4L);
-		BILL_SUBTYPE_MAP.put("Supplies", 5L);
-		BILL_SUBTYPE_MAP.put("Recovery", 6L);
-		BILL_SUBTYPE_MAP.put("Deposit", 7L);
-		BILL_SUBTYPE_MAP.put("Advance", 8L);
-		BILL_SUBTYPE_MAP.put("GPF", 9L);
-		BILL_SUBTYPE_MAP.put("Others", 10L);
-		BILL_SUBTYPE_MAP.put("Expense", 11L);
+		BILL_SUBTYPE_MAP.put("Contingent", 1);
+		BILL_SUBTYPE_MAP.put("Salary", 2);
+		BILL_SUBTYPE_MAP.put("Pension", 3);
+		BILL_SUBTYPE_MAP.put("Works", 4);
+		BILL_SUBTYPE_MAP.put("Supplies", 5);
+		BILL_SUBTYPE_MAP.put("Recovery", 6);
+		BILL_SUBTYPE_MAP.put("Deposit", 7);
+		BILL_SUBTYPE_MAP.put("Advance", 8);
+		BILL_SUBTYPE_MAP.put("GPF", 9);
+		BILL_SUBTYPE_MAP.put("Others", 10);
+		BILL_SUBTYPE_MAP.put("Expense", 11);
 	}
 
-	public static Long getBillSubTypeId(String billSubtype) {
+	public static Integer getBillSubTypeId(String billSubtype) {
 
 		if (billSubtype == null || billSubtype.trim().isEmpty()) {
 			return null;
